Bombers and Knights oh my

+++ games/TurnBasedGames/Bouncy/js/GameElements/Projectiles/SingleHitProjectile.js
@@ -12,6 +12,8 @@ class SingleHitProjectile extends Projectile {
       EffectFactory.createDamageEffect(boardState, intersection);
     }
 
-    this.delete();
+    if (!intersection.line.forceBounce()) {
+      this.delete();
+    }
   }
 }
